Added a sub-package for "creational" in order to organize the code better
