+++ src/main/java/com/designpatterns/creational/factory/Pentagon.java
@@ -1,4 +1,4 @@
-package src.main.java.com.designpatterns.factorypattern;
+package src.main.java.com.designpatterns.creational.factory;
 
 public class Pentagon implements Polygon {
     @Override
